Separate content from ScrollingStack logic
