+++ packages/demo/frontend/tailwind.config.js
@@ -11,6 +11,7 @@ export default {
         orange: '#fe8019',
         yellow: '#fabd2f',
         green: '#b8bb26',
+        aqua: '#8ec07c',
         blue: '#83a598',
         indigo: '#d3869b',
         purple: '#d3869b',
